requirements.txt
scikit-learn==1.5.0
pandas>=2.2
pyyaml>=6.0
joblib>=1.4
